requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
httpx>=0.26.0
rank-bm25>=0.2.2
tiktoken>=0.5.2
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
pytest
